weaviate_loader/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.9
    # via llama-index-core
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via httpx
attrs==24.2.0
    # via
    #   aiohttp
    #   environ-config
authlib==1.3.1
    # via weaviate-client
beautifulsoup4==4.12.3
    # via llama-index-readers-file
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
cryptography==43.0.1
    # via authlib
dataclasses-json==0.6.7
    # via llama-index-core
deprecated==1.2.14
    # via llama-index-core
dirtyjson==1.0.8
    # via llama-index-core
environ-config==24.1.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via llama-index-core
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.66.2
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.66.2
    # via weaviate-client
grpcio-tools==1.66.2
    # via weaviate-client
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.0
    # via
    #   -r requirements.in
    #   llama-index-core
    #   weaviate-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
joblib==1.4.2
    # via nltk
llama-index-core==0.11.16
    # via
    #   -r requirements.in
    #   llama-index-readers-file
llama-index-readers-file==0.2.2
    # via -r requirements.in
marshmallow==3.22.0
    # via dataclasses-json
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via llama-index-core
networkx==3.3
    # via llama-index-core
nltk==3.9.1
    # via llama-index-core
numpy==1.26.4
    # via
    #   llama-index-core
    #   pandas
packaging==24.1
    # via marshmallow
pandas==2.2.3
    # via
    #   -r requirements.in
    #   llama-index-readers-file
pillow==10.4.0
    # via llama-index-core
protobuf==5.28.2
    # via
    #   grpcio-health-checking
    #   grpcio-tools
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   llama-index-core
    #   weaviate-client
pydantic-core==2.23.4
    # via pydantic
pypdf==4.3.1
    # via llama-index-readers-file
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via llama-index-core
regex==2024.9.11
    # via
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   llama-index-core
    #   tiktoken
    #   weaviate-client
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy[asyncio]==2.0.35
    # via
    #   llama-index-core
    #   sqlalchemy
striprtf==0.0.26
    # via llama-index-readers-file
tenacity==8.5.0
    # via llama-index-core
tiktoken==0.8.0
    # via llama-index-core
tqdm==4.66.5
    # via
    #   llama-index-core
    #   nltk
typing-extensions==4.12.2
    # via
    #   llama-index-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
validators==0.34.0
    # via weaviate-client
weaviate-client==4.8.1
    # via -r requirements.in
wrapt==1.16.0
    # via
    #   deprecated
    #   llama-index-core
yarl==1.13.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
